+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2023-11-27
+### Added
+- Added `interface(T)` macro attribute to provide default interface bindings for the component builders, allowing to cut down verboseness of catalog configuration. Example:
+  ```rust
+  trait Iface {}
+
+  #[dill::component]
+  #[dill::interface(dyn Iface)]
+  struct Impl;
+
+  // Automatically does `.bind::<dyn Iface, Impl>()`
+  CatalogBuilder::new().add::<Impl>().build();
+  ```
+
 ## [0.7.2] - 2023-09-04
 ### Changed
 - Fixed validation of Catalog self-injection